requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langdetect
transformers
pydantic
sqlalchemy
psycopg2-binary
